requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
django-picklefield==3.0.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.6
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
MarkupSafe==2.0.1
mongoengine==0.23.1
packaging==21.3
pymongo==3.12.3
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.2.4
uritemplate==4.1.1
urllib3==1.26.7
